modified: pass cc 7 assignments to keyboardcontroller for velocity

+++ firmware/src/objects/Settings.h
@@ -23,8 +23,8 @@ struct LeverPushSettings {
     int minCCValue;
     int maxCCValue;
     LeverPushFunctionMode functionMode;
-    int onsetTime;
-    int offsetTime;
+    unsigned long onsetTime;
+    unsigned long offsetTime;
     InterpolationType onsetType;
     InterpolationType offsetType;
 };
